requirements.txt
chromadb==0.4.24
docx==0.2.4
fastapi==0.111.0
fitz==0.0.1.dev2
jose==1.0.0
nltk==3.8.1
pydantic==1.10.2
python_bcrypt==0.3.2
python_docx==1.1.2
python_jose==3.3.0
SQLAlchemy==2.0.27
uvicorn==0.29.0
